fix01

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: marvin@42.fr <astachni>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 16:06:25 by astachni          #+#    #+#             */
-/*   Updated: 2022/11/10 15:42:15 by marvin@42.f      ###   ########.fr       */
+/*   Updated: 2022/11/14 20:50:37 by marvin@42.f      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
 	int	i;
